+++ src/tables/mod.rs
@@ -28,3 +28,4 @@ pub mod post;
 pub mod stat;
 pub mod vhea;
 pub mod vmtx;
+pub mod xvar;